:recycle: refactor `register` module
